Remove unused local state structs
