+++ src/material/mod.rs
@@ -3,10 +3,12 @@
 mod conductivity;
 mod linear_elastic;
 mod local_state;
+mod local_state_porous;
 mod stress_strain;
 mod von_mises;
 pub use crate::material::conductivity::*;
 pub use crate::material::linear_elastic::*;
 pub use crate::material::local_state::*;
+pub use crate::material::local_state_porous::*;
 pub use crate::material::stress_strain::*;
 pub use crate::material::von_mises::*;
